Fix: Рефакторит типы
