+++ src/components/map/map.tsx
@@ -1,6 +1,6 @@
 import {useRef, useEffect} from 'react';
 import {Icon, Marker, layerGroup} from 'leaflet';
-import {TMapCity, TMapPoint} from '../../utils/types';
+import {TMapCity, TMapPoint} from '../../types/map';
 import {MapType} from '../../utils/const';
 import useMap from '../../hooks/use-map';
 import 'leaflet/dist/leaflet.css';
