+++ src/components/offer-list/offer-list.tsx
@@ -1,4 +1,4 @@
-import {TCityName, TOfferPreview} from '../../utils/types';
+import {TCityName, TOfferPreview} from '../../types/offers';
 import OffersSortingPanel from '../offers-sorting-panel';
 import OfferCardList from '../offer-card-list';
 import {OfferCardParams} from '../../utils/const';
