+++ src/components/offer-reviews-item/offer-reviews-item.tsx
@@ -1,6 +1,6 @@
 import dayjs from 'dayjs';
 import {REVIEW_DATE_FORMAT, RatingPanelType} from '../../utils/const';
-import {TReview} from '../../utils/types';
+import {TReview} from '../../types/reviews';
 import RatingPanel from '../rating-panel';
 
 type TOfferReviewsItemProps = {
