+++ src/components/offer-reviews-list/offer-reviews-list.tsx
@@ -1,4 +1,4 @@
-import {TReview} from '../../utils/types';
+import {TReview} from '../../types/reviews';
 import OfferReviewsItem from '../offer-reviews-item';
 
 type TOfferReviewsListProps = {
