+++ src/hooks/use-map.ts
@@ -1,6 +1,6 @@
 import {useEffect, useState, MutableRefObject, useRef} from 'react';
 import {Map, TileLayer} from 'leaflet';
-import {TMapCity} from '../utils/types';
+import {TMapCity} from '../types/map';
 
 const DEFAULT_ZOOM = 10;
 const URL_TEMPLATE = 'https://{s}.basemaps.cartocdn.com/rastertiles/voyager/{z}/{x}/{y}{r}.png';
