+++ src/mocks/offer.ts
@@ -1,4 +1,4 @@
-import {TOffer} from '../utils/types';
+import {TOffer} from '../types/offers';
 
 export const offer: TOffer = {
   id: '843afec7-4e33-430f-8042-cd2d928c95a0',
