+++ src/store/app-data/selectors.ts
@@ -1,5 +1,5 @@
 import {NameSpace} from '../../utils/const';
-import {State} from '../../utils/state';
-import {TOfferPreview} from '../../utils/types';
+import {State} from '../../types/state';
+import {TOfferPreview} from '../../types/offers';
 
 export const getOffers = (state: State): TOfferPreview[] => state[NameSpace.Data].offers;
