+++ src/store/app-process/selectors.ts
@@ -1,6 +1,6 @@
 import {NameSpace} from '../../utils/const';
-import {State} from '../../utils/state';
-import {TCityName} from '../../utils/types';
+import {State} from '../../types/state';
+import {TCityName} from '../../types/offers';
 
 export const getCurrentCity = (state: Pick<State, NameSpace.App>): TCityName => state[NameSpace.App].currentCity;
 export const getCurrentOffer = (state: Pick<State, NameSpace.App>): string | null => state[NameSpace.App].currentOffer;
