+++ src/types/event-handlers.ts
@@ -1,0 +1,5 @@
+import {ChangeEventHandler, MouseEventHandler} from 'react';
+
+export type TInputChangeHandler = ChangeEventHandler<HTMLInputElement>;
+export type TFormChangeHandler = ChangeEventHandler<HTMLInputElement | HTMLTextAreaElement>;
+export type TSpanClickHandler = MouseEventHandler<HTMLSpanElement>;
