+++ src/types/map.ts
@@ -1,0 +1,13 @@
+export type TMapCity = {
+  title: string;
+  lat: number;
+  lng: number;
+  zoom: number;
+};
+
+export type TMapPoint = {
+  id: string;
+  title: string;
+  lat: number;
+  lng: number;
+};
