--- src/utils/state.ts
@@ -1,4 +1,0 @@
-import {store} from '../store';
-
-export type State = ReturnType<typeof store.getState>;
-export type AppDispatch = typeof store.dispatch;
